@@ -10,9 +10,7 @@ import org.lwjgl.opengl.GL30C;
 public class GlVertexArray extends GlObject {
     public static final int NULL_ARRAY_ID = 0;
 
-    public GlVertexArray(RenderDevice owner) {
-        super(owner);
-
+    public GlVertexArray() {
         this.setHandle(GL30C.glGenVertexArrays());
     }
 }
