@@ -1,7 +1,5 @@
 package me.jellysquid.mods.sodium.client.gl.buffer;
 
-import me.jellysquid.mods.sodium.client.gl.device.RenderDevice;
-
 /**
  * A mutable buffer type which is supported with OpenGL 1.5+. The buffer's storage can be reallocated at any time
  * without needing to re-create the buffer itself.
@@ -9,8 +7,8 @@ import me.jellysquid.mods.sodium.client.gl.device.RenderDevice;
 public class GlMutableBuffer extends GlBuffer {
     private long size = 0L;
 
-    public GlMutableBuffer(RenderDevice owner, GlBufferUsage usage) {
-        super(owner, usage);
+    public GlMutableBuffer() {
+        super();
     }
 
     public void setSize(long size) {
