@@ -1,21 +1,16 @@
 package me.jellysquid.mods.sodium.client.gl.buffer;
 
 import me.jellysquid.mods.sodium.client.gl.attribute.GlVertexFormat;
-
-import java.nio.ByteBuffer;
+import me.jellysquid.mods.sodium.client.util.NativeBuffer;
 
 /**
  * Helper type for tagging the vertex format alongside the raw buffer data.
  */
-public class IndexedVertexData {
-    public final GlVertexFormat<?> vertexFormat;
-
-    public final ByteBuffer vertexBuffer;
-    public final ByteBuffer indexBuffer;
-
-    public IndexedVertexData(GlVertexFormat<?> vertexFormat, ByteBuffer vertexBuffer, ByteBuffer indexBuffer) {
-        this.vertexFormat = vertexFormat;
-        this.vertexBuffer = vertexBuffer;
-        this.indexBuffer = indexBuffer;
+public record IndexedVertexData(GlVertexFormat<?> vertexFormat,
+                                NativeBuffer vertexBuffer,
+                                NativeBuffer indexBuffer) {
+    public void delete() {
+        this.vertexBuffer.free();
+        this.indexBuffer.free();
     }
 }
