@@ -1,20 +1,14 @@
 package me.jellysquid.mods.sodium.client.gl.device;
 
 import me.jellysquid.mods.sodium.client.gl.array.GlVertexArray;
-import me.jellysquid.mods.sodium.client.gl.buffer.GlBuffer;
-import me.jellysquid.mods.sodium.client.gl.buffer.GlBufferTarget;
-import me.jellysquid.mods.sodium.client.gl.buffer.GlBufferUsage;
-import me.jellysquid.mods.sodium.client.gl.buffer.GlMutableBuffer;
+import me.jellysquid.mods.sodium.client.gl.buffer.*;
 import me.jellysquid.mods.sodium.client.gl.state.GlStateTracker;
 import me.jellysquid.mods.sodium.client.gl.tessellation.GlPrimitiveType;
 import me.jellysquid.mods.sodium.client.gl.tessellation.GlTessellation;
 import me.jellysquid.mods.sodium.client.gl.tessellation.GlVertexArrayTessellation;
 import me.jellysquid.mods.sodium.client.gl.tessellation.TessellationBinding;
 import org.lwjgl.PointerBuffer;
-import org.lwjgl.opengl.GL20C;
-import org.lwjgl.opengl.GL30C;
-import org.lwjgl.opengl.GL31C;
-import org.lwjgl.opengl.GL32C;
+import org.lwjgl.opengl.*;
 
 import java.nio.ByteBuffer;
 import java.nio.IntBuffer;
@@ -75,12 +69,16 @@ public class GLRenderDevice implements RenderDevice {
         }
 
         @Override
-        public void uploadData(GlMutableBuffer glBuffer, ByteBuffer byteBuffer) {
+        public void uploadData(GlMutableBuffer glBuffer, ByteBuffer byteBuffer, GlBufferUsage usage) {
             this.bindBuffer(GlBufferTarget.ARRAY_BUFFER, glBuffer);
 
-            GL20C.glBufferData(GlBufferTarget.ARRAY_BUFFER.getTargetParameter(), byteBuffer, glBuffer.getUsageHint().getId());
-
-            glBuffer.setSize(byteBuffer.limit());
+            if (byteBuffer == null) {
+                GL20C.glBufferData(GlBufferTarget.ARRAY_BUFFER.getTargetParameter(), 0L, usage.getId());
+                glBuffer.setSize(0L);
+            } else {
+                GL20C.glBufferData(GlBufferTarget.ARRAY_BUFFER.getTargetParameter(), byteBuffer, usage.getId());
+                glBuffer.setSize(byteBuffer.remaining());
+            }
         }
 
         @Override
@@ -110,20 +108,17 @@ public class GLRenderDevice implements RenderDevice {
         }
 
         @Override
-        public void invalidateBuffer(GlMutableBuffer glBuffer) {
-            this.allocateBuffer(GlBufferTarget.ARRAY_BUFFER, glBuffer, 0L);
-        }
+        public void allocateStorage(GlMutableBuffer buffer, long bufferSize, GlBufferUsage usage) {
+            this.bindBuffer(GlBufferTarget.ARRAY_BUFFER, buffer);
 
-        @Override
-        public void allocateBuffer(GlBufferTarget target, GlMutableBuffer buffer, long bufferSize) {
-            this.bindBuffer(target, buffer);
-
-            GL20C.glBufferData(target.getTargetParameter(), bufferSize, buffer.getUsageHint().getId());
+            GL20C.glBufferData(GlBufferTarget.ARRAY_BUFFER.getTargetParameter(), bufferSize, usage.getId());
             buffer.setSize(bufferSize);
         }
 
         @Override
         public void deleteBuffer(GlBuffer buffer) {
+            this.stateTracker.notifyBufferDeleted(buffer);
+
             int handle = buffer.handle();
             buffer.invalidateHandle();
 
@@ -131,9 +126,11 @@ public class GLRenderDevice implements RenderDevice {
         }
 
         @Override
-        public void deleteVertexArray(GlVertexArray array) {
-            int handle = array.handle();
-            array.invalidateHandle();
+        public void deleteVertexArray(GlVertexArray vertexArray) {
+            this.stateTracker.notifyVertexArrayDeleted(vertexArray);
+
+            int handle = vertexArray.handle();
+            vertexArray.invalidateHandle();
 
             GL30C.glDeleteVertexArrays(handle);
         }
@@ -157,13 +154,13 @@ public class GLRenderDevice implements RenderDevice {
         }
 
         @Override
-        public GlMutableBuffer createMutableBuffer(GlBufferUsage usage) {
-            return new GlMutableBuffer(GLRenderDevice.this, usage);
+        public GlMutableBuffer createMutableBuffer() {
+            return new GlMutableBuffer();
         }
 
         @Override
         public GlTessellation createTessellation(GlPrimitiveType primitiveType, TessellationBinding[] bindings, GlBuffer indexBuffer) {
-            GlVertexArrayTessellation tessellation = new GlVertexArrayTessellation(new GlVertexArray(GLRenderDevice.this), primitiveType, bindings, indexBuffer);
+            GlVertexArrayTessellation tessellation = new GlVertexArrayTessellation(new GlVertexArray(), primitiveType, bindings, indexBuffer);
             tessellation.init(this);
 
             return tessellation;
@@ -179,12 +176,6 @@ public class GLRenderDevice implements RenderDevice {
         public void multiDrawElementsBaseVertex(PointerBuffer pointer, IntBuffer count, IntBuffer baseVertex) {
             GlPrimitiveType primitiveType = GLRenderDevice.this.activeTessellation.getPrimitiveType();
             GL32C.glMultiDrawElementsBaseVertex(primitiveType.getId(), count, GL20C.GL_UNSIGNED_INT, pointer, baseVertex);
-        }
-
-        @Override
-        public void multiDrawElements(PointerBuffer pointer, IntBuffer count) {
-            GlPrimitiveType primitiveType = GLRenderDevice.this.activeTessellation.getPrimitiveType();
-            GL32C.glMultiDrawElements(primitiveType.getId(), count, GL20C.GL_UNSIGNED_INT, pointer);
         }
 
         @Override
