@@ -15,9 +15,7 @@ public class GlShader extends GlObject {
 
     private final Identifier name;
 
-    public GlShader(RenderDevice owner, ShaderType type, Identifier name, String src) {
-        super(owner);
-
+    public GlShader(ShaderType type, Identifier name, String src) {
         this.name = name;
 
         int handle = GL20C.glCreateShader(type.id);
