@@ -37,10 +37,12 @@ import java.util.Map;
 public class ChunkRenderRebuildTask extends ChunkRenderBuildTask {
     private final RenderSection render;
     private final ChunkRenderContext context;
+    private final int frame;
 
-    public ChunkRenderRebuildTask(RenderSection render, ChunkRenderContext context) {
+    public ChunkRenderRebuildTask(RenderSection render, ChunkRenderContext context, int frame) {
         this.render = render;
         this.context = context;
+        this.frame = frame;
     }
 
     @Override
@@ -153,7 +155,7 @@ public class ChunkRenderRebuildTask extends ChunkRenderBuildTask {
         renderData.setOcclusionData(occluder.build());
         renderData.setBounds(bounds.build(this.render.getChunkPos()));
 
-        return new ChunkBuildResult(this.render, renderData.build(), meshes);
+        return new ChunkBuildResult(this.render, renderData.build(), meshes, this.frame);
     }
 
     @Override
