@@ -1,34 +1,6 @@
 package me.jellysquid.mods.sodium.client.util;
 
-public class Dim2i {
-    private final int x;
-    private final int y;
-    private final int width;
-    private final int height;
-
-    public Dim2i(int x, int y, int width, int height) {
-        this.x = x;
-        this.y = y;
-        this.width = width;
-        this.height = height;
-    }
-
-    public int getOriginX() {
-        return this.x;
-    }
-
-    public int getOriginY() {
-        return this.y;
-    }
-
-    public int getWidth() {
-        return this.width;
-    }
-
-    public int getHeight() {
-        return this.height;
-    }
-
+public record Dim2i(int x, int y, int width, int height) {
     public int getLimitX() {
         return this.x + this.width;
     }
